plant_disease_app/requirements.txt
%%writefile /content/plant_disease_app/requirements.txt
flask
torch
torchvision
transformers
Pillow
scikit-learn
requests
